+++ Sources/SpotAsk/Settings/ConfigBackup.swift
@@ -79,6 +79,7 @@ struct SpotAskConfigBackup: Codable, Equatable, Sendable {
         var hotKeyPreset: String
         var keepWindowOnTop: Bool
         var showsMenuBarIcon: Bool
+        var automaticUpdateCheckEnabled: Bool?
         var proxyEnabled: Bool?
         var proxyType: String?
         var proxyHost: String?